add filter_bed utils to vcfbed package
